requirements.txt
pyyaml
jsonschema
watchdog
Flask
jsonpatch
